+++ core/rhi/rorvertex_attribute.hpp
@@ -61,7 +61,7 @@ class ROAR_ENGINE_ITEM VertexAttribute final
 	 */
 	FORCE_INLINE VertexAttribute(uint32_t            a_location,
 								 uint32_t            a_offset,
-								 uint64_t            a_buffer_offset,
+								 ptrdiff_t           a_buffer_offset,
 								 uint32_t            a_binding,
 								 uint32_t            a_buffer_index,
 								 rhi::BufferSemantic a_semantics = rhi::BufferSemantic::vertex_position,
@@ -79,7 +79,7 @@ class ROAR_ENGINE_ITEM VertexAttribute final
 	// clang-format off
 	FORCE_INLINE uint32_t            location()         const { return this->m_location;         }
 	FORCE_INLINE uint32_t            offset()           const { return this->m_offset;           }
-	FORCE_INLINE uint64_t            buffer_offset()    const { return this->m_buffer_offset;    }
+	FORCE_INLINE ptrdiff_t           buffer_offset()    const { return this->m_buffer_offset;    }
 	FORCE_INLINE uint32_t            binding()          const { return this->m_binding;          }
 	FORCE_INLINE uint32_t            buffer_index()     const { return this->m_buffer_index;     }
 	FORCE_INLINE rhi::VertexFormat   format()           const { return this->m_format;           }
@@ -88,7 +88,7 @@ class ROAR_ENGINE_ITEM VertexAttribute final
 
 	FORCE_INLINE void   location(uint32_t a_location)              noexcept { this->m_location         = a_location;     }
 	FORCE_INLINE void   offset(uint32_t a_offset)                  noexcept { this->m_offset           = a_offset;       }
-	FORCE_INLINE void   buffer_offset(uint64_t a_offset)           noexcept { this->m_buffer_offset    = a_offset;       }
+	FORCE_INLINE void   buffer_offset(ptrdiff_t a_offset)          noexcept { this->m_buffer_offset    = a_offset;       }
 	FORCE_INLINE void   binding(uint32_t a_binding)                noexcept { this->m_binding          = a_binding;      }
 	FORCE_INLINE void   buffer_index(uint32_t a_index)             noexcept { this->m_buffer_index     = a_index;        }
 	FORCE_INLINE void   format(rhi::VertexFormat a_format)         noexcept { this->m_format           = a_format;       }
@@ -99,7 +99,7 @@ class ROAR_ENGINE_ITEM VertexAttribute final
   private:
 	uint32_t            m_location{0};                                            //! Use in shader like "layout(location = m_location)", decided by layout
 	uint32_t            m_offset{0};                                              //! Offset of this attribute relative to other attributes
-	uint64_t            m_buffer_offset{0};                                       //! Offset of this attribute in buffer buffer_index from the start, decided by layout
+	ptrdiff_t           m_buffer_offset{0};                                       //! Offset of this attribute in buffer buffer_index from the start, decided by layout
 	uint32_t            m_binding{0};                                             //! Which layout binding describes this attribute
 	uint32_t            m_buffer_index{0};                                        //! Which buffer do I live in? this is the buffer_pack index, destination buffer index
 	rhi::BufferSemantic m_semantics{rhi::BufferSemantic::vertex_position};        //! Whats the type of attribute? Position, UV, Normal etc
